move CommandRunner to its own file
